rename lgr to xlog package
