+++ internal/handler/rest/auth.go
@@ -6,8 +6,8 @@ import (
 	"strings"
 
 	"github.com/jonashiltl/openchangelog/internal/errs"
-	"github.com/jonashiltl/openchangelog/internal/lgr"
 	"github.com/jonashiltl/openchangelog/internal/store"
+	"github.com/jonashiltl/openchangelog/internal/xlog"
 )
 
 type Token struct {
@@ -30,7 +30,7 @@ func bearerAuth(e *env, r *http.Request) (Token, error) {
 	if err != nil {
 		return Token{}, err
 	}
-	lgr.AddWorkspaceID(r, id.String())
+	xlog.AddWorkspaceID(r, id.String())
 	return Token{
 		Key:         key,
 		WorkspaceID: id,
