+++ internal/handler/rss/routes.go
@@ -7,9 +7,9 @@ import (
 
 	"github.com/jonashiltl/openchangelog/internal/config"
 	"github.com/jonashiltl/openchangelog/internal/errs"
-	"github.com/jonashiltl/openchangelog/internal/lgr"
 	"github.com/jonashiltl/openchangelog/internal/load"
 	"github.com/jonashiltl/openchangelog/internal/parse"
+	"github.com/jonashiltl/openchangelog/internal/xlog"
 )
 
 type env struct {
@@ -31,7 +31,7 @@ func RegisterRSSHandler(mux *http.ServeMux, e *env) {
 }
 
 func serveHTTP(env *env, h func(e *env, w http.ResponseWriter, r *http.Request) error) http.HandlerFunc {
-	return lgr.AttachLogger(func(w http.ResponseWriter, r *http.Request) {
+	return xlog.AttachLogger(func(w http.ResponseWriter, r *http.Request) {
 		err := h(env, w, r)
 		if err != nil {
 			status := http.StatusInternalServerError
@@ -69,7 +69,7 @@ func serveHTTP(env *env, h func(e *env, w http.ResponseWriter, r *http.Request) 
 				http.Error(w, err.Error(), http.StatusInternalServerError)
 			}
 
-			lgr.LogRequest(r.Context(), status, msg)
+			xlog.LogRequest(r.Context(), status, msg)
 		}
 	})
 }
