+++ internal/xlog/log.go
@@ -1,4 +1,4 @@
-package lgr
+package xlog
 
 import (
 	"context"
